Implemented the snippetForm modal.
Can only edit or add snippets one at a time. All other components on
the page are disabled due to a modal div. This made the code so much
easier and safe. The header is still alive, but the search is disabled
during snippet add or edit. This took a LOT of work!