Removed venv directory from .gitignore
This will store all of the virtual env, not just parts that are there
now. That happened when I  put the ‘venv’ tag in .gitignore. It stopped
pushing any venv update into GitHub. I’m not real sure what to do about
this - should venv be there or not. I guess I could always just
recreate the venv each time. But does it need to be pushed into the
heroku cloud?